+++ EnhanceQoL/EnhanceQoL.lua
@@ -2367,15 +2367,15 @@ local function initUI()
 			LDBIcon:Hide(addonName)
 		end
 	end
-	function addon.functions.toggleZoneText(value)
+	function addon.functions.toggleZoneText(value, ignore)
 		if value then
 			ZoneTextFrame:UnregisterAllEvents()
 			ZoneTextFrame:Hide()
-		else
+		elseif not ignore then
 			addon.variables.requireReload = true
 		end
 	end
-	addon.functions.toggleZoneText(addon.db["hideZoneText"])
+	addon.functions.toggleZoneText(addon.db["hideZoneText"], true)
 
 	function addon.functions.toggleQuickJoinToastButton(value)
 		if value == false then
@@ -3367,7 +3367,13 @@ local function CreateUI()
 		root:CreateButton(L["SettingsDataPanelCreate"], function() local dialog = StaticPopup_Show("EQOL_CREATE_DATAPANEL") end)
 
 		DoDevider()
-		root:CreateButton(SETTINGS, function() Settings.OpenToCategory(addon.SettingsLayout.rootCategory:GetID()) end)
+		root:CreateButton(LFG_LIST_LEGACY .. " " .. SETTINGS, function()
+			if frame:IsShown() then
+				frame:Hide()
+			else
+				frame:Show()
+			end
+		end)
 	end
 
 	-- Datenobjekt fr den Minimap-Button
@@ -3377,11 +3383,7 @@ local function CreateUI()
 		icon = "Interface\\AddOns\\" .. addonName .. "\\Icons\\Icon.tga", -- Hier kannst du dein eigenes Icon verwenden
 		OnClick = function(_, msg)
 			if msg == "LeftButton" then
-				if frame:IsShown() then
-					frame:Hide()
-				else
-					frame:Show()
-				end
+				Settings.OpenToCategory(addon.SettingsLayout.rootCategory:GetID())
 			else
 				MenuUtil.CreateContextMenu(UIParent, QuickMenuGenerator)
 			end