+++ EnhanceQoL/NewSettingsTable.lua
@@ -2,6 +2,9 @@ local addonName, addon = ...
 
 addon.variables.NewVersionTableEQOL = {
 	-- Headlines
+	["CombatDungeons"] = true,
+	["EQOL_WorldMarkerCycle"] = true,
+
 	["UIInput"] = true,
 	["EQOL_modifyXPRepBar"] = true,
 	["EQOL_modifyXPRepBarWidth"] = true,